move metrics to utils submodule
