+++ responsibleai_text/responsibleai_text/utils/genai_metrics/scripts/coherence.py
@@ -10,7 +10,7 @@ logger = evaluate.logging.get_logger(__name__)
 _CITATION = """
 """
 
-_DESCRIPTION = """The relevance metric.
+_DESCRIPTION = """The coherence metric.
 """
 
 _KWARGS_DESCRIPTION = """
@@ -23,28 +23,36 @@ Your response will be used in automated evaluation of question-answering systems
 """.strip()
 
 _TEMPLATE = """
-Relevance measures how well the answer addresses the main aspects of the question, based on the context. Consider whether all and only the important aspects are contained in the answer when evaluating relevance. Given the context and question, score the relevance of the answer between one to five stars using the following rating scale:
-One star: the answer completely lacks relevance
-Two stars: the answer mostly lacks relevance
-Three stars: the answer is partially relevant
-Four stars: the answer is mostly relevant
-Five stars: the answer has perfect relevance
+Coherence of an answer is measured by how well all the sentences fit together and sound naturally as a whole. Consider the overall quality of the answer when evaluating coherence. Given the question and answer, score the coherence of answer between one to five stars using the following rating scale:
+One star: the answer completely lacks coherence
+Two stars: the answer mostly lacks coherence
+Three stars: the answer is partially coherent
+Four stars: the answer is mostly coherent
+Five stars: the answer has perfect coherency
 
 This rating value should always be an integer between 1 and 5. So the rating produced should be 1 or 2 or 3 or 4 or 5.
-
-CONTEXT:
-{context}
+Some examples of valid responses are:
+1
+2
+5
+Some examples of invalid responses are:
+1/5
+1.5
+3.0
+5 stars
 
 QUESTION:
 {question}
 
 ANSWER:
 {prediction}
+
+RATING:
 """.strip()
 
 
 @evaluate.utils.file_utils.add_start_docstrings(_DESCRIPTION, _KWARGS_DESCRIPTION)
-class Relevance(evaluate.Metric):
+class Coherence(evaluate.Metric):
     def _info(self):
 
         return evaluate.MetricInfo(
@@ -54,8 +62,7 @@ class Relevance(evaluate.Metric):
             features=datasets.Features(
                 {
                     "predictions": datasets.Value("string", id="sequence"),
-                    "references": datasets.Value("string", id="sequence"),
-                    "questions": datasets.Value("string", id="sequence")
+                    "references": datasets.Value("string", id="sequence")
                 }
             ),
         )
@@ -64,9 +71,8 @@ class Relevance(evaluate.Metric):
         m = []
         templated_ques = []
 
-        questions = kwargs['questions']
-        for p, r, q in zip(predictions, references, questions):
-            templated_ques.append(_TEMPLATE.format(context=r, question=q, prediction=p))
+        for p, r in zip(predictions, references):
+            templated_ques.append(_TEMPLATE.format(question=r, prediction=p))
 
         model = kwargs['wrapper_model']
